Add support for QoS 1 in client
It involves storing QoS 1 PUBLISH packets in the session until they are
acknowledged by the server (with PUBACK) and resending them when reconnecting
if the server indicates it has a session present.

According to specification, the session is cleared if the server
indicates it has no session present. In that case, messages are not
republished.